# Conflicts:
#	ApplicationsXML/application.xml
#	Source/Relativity.IntegrationPoints.Tests/Integration/Helpers/RelativityHelpers/JobBuilder.cs
#	Source/Relativity.IntegrationPoints.Tests/Integration/Helpers/RelativityHelpers/JobHelper.cs
#	Source/Relativity.IntegrationPoints.Tests/Integration/Helpers/WorkspaceHelpers/FieldsMappingHelper.cs
#	Source/Relativity.IntegrationPoints.Tests/Integration/Mocks/FakeAgent.cs
#	Source/Relativity.IntegrationPoints.Tests/Integration/Tests/Agent/ImportServiceManagerTests.cs
#	Source/Relativity.IntegrationPoints.Tests/Integration/Tests/Agent/SyncWorkerLDAPProviderTests.cs
#	Source/Relativity.IntegrationPoints.Tests/Integration/Tests/Agent/SyncWorkerMyFirstProviderTests.cs
#	Source/Relativity.IntegrationPoints.Tests/Integration/Tests/ScheduleQueue/ScheduleAgentTests.cs
#	Source/Relativity.IntegrationPoints.Tests/Integration/TestsBase.cs

+++ Source/kCura.IntegrationPoints.Core.Contracts/Import/LoadFileTaskParameters.cs
@@ -5,7 +5,7 @@ namespace kCura.IntegrationPoints.Core.Contracts.Import
 	[Serializable]
 	public class LoadFileTaskParameters
 	{
-		public long ProcessedItemsCount { get; set; }
+		public int ProcessedItemsCount { get; set; }
 
 		public long Size { get; set; }
